requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==41.0.7
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
idna==3.10
mysqlclient==2.2.7
pycparser==2.22
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
